requirements.txt
# Core Dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0

# AI/ML Libraries
google-generativeai>=0.3.2
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
duckduckgo-search>=6.3.5

# Vector Stores & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
fastembed>=0.2.4
rank-bm25>=0.2.2
tqdm>=4.65.0

# Document Processing
pypdf>=3.17.1
PyPDF2>=3.0.1

# Web Search (already included above)

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Data Handling
pandas>=2.1.1
numpy>=1.26.0
